initial nullsafety upgrade

+++ lib/src/popup_position.dart
@@ -70,7 +70,7 @@ class PopupPosition {
   ) {
     return PopupContainer(
       size: visibleSize,
-      right: visibleSize.x - markerPosition.x,
+      right: visibleSize.x - (markerPosition.x as double),
       top: markerPosition.y - (visibleSize.y / 2) + (markerSize.y / 2),
       alignment: Alignment.centerRight,
     );
@@ -85,7 +85,7 @@ class PopupPosition {
       size: visibleSize,
       alignment: Alignment.bottomCenter,
       left: markerPosition.x - (visibleSize.x / 2) + (markerSize.x / 2),
-      bottom: visibleSize.y - markerPosition.y,
+      bottom: visibleSize.y - (markerPosition.y as double),
     );
   }
 
@@ -97,7 +97,7 @@ class PopupPosition {
     return PopupContainer(
       size: visibleSize,
       alignment: Alignment.centerLeft,
-      left: markerPosition.x + markerSize.x,
+      left: markerPosition.x + (markerSize.x as double),
       top: markerPosition.y - (visibleSize.y / 2) + (markerSize.y / 2),
     );
   }
@@ -111,7 +111,7 @@ class PopupPosition {
       size: visibleSize,
       alignment: Alignment.topCenter,
       left: markerPosition.x - (visibleSize.x / 2) + (markerSize.x / 2),
-      top: markerPosition.y + markerSize.y,
+      top: markerPosition.y + (markerSize.y as double),
     );
   }
